created firebase.js file and changed impo statements
